+++ docs/source/reference/backends/macaulay2.rst
@@ -1,6 +1,6 @@
 ``macaulay2``
 =============
 
-.. automodule:: sage_acsv.macaulay2
+.. automodule:: sage_acsv.backends.macaulay2
    :members:
    :undoc-members:
