+++ docs/source/reference/backends/msolve.rst
@@ -1,6 +1,6 @@
 ``msolve``
 ==========
 
-.. automodule:: sage_acsv.msolve
+.. automodule:: sage_acsv.backends.msolve
    :members:
    :undoc-members:
