+++ sage_acsv/settings.py
@@ -44,13 +44,15 @@ class KroneckerBackend(Enum):
 
 class GroebnerBackend(Enum):
     """Options for computing Groebner Bases and related ideal functions
-    :func:`.compute_groebner_basis`, :func:`.compute_primary_decomposition`, and :func:`.critical_points`.
+    :func:`.groebner.compute_groebner_basis`,
+    :func:`.groebner.compute_primary_decomposition`,
+    and :func:`.groebner.critical_points`.
 
     See also:
 
-    - :func:`.compute_groebner_basis`
-    - :func:`.compute_primary_decomposition`
-    - :func:`.compute_radical`
+    - :func:`.groebner.compute_groebner_basis`
+    - :func:`.groebner.compute_primary_decomposition`
+    - :func:`.groebner.compute_radical`
     - :class:`.ACSVSettings`
 
     """
@@ -180,7 +182,9 @@ class ACSVSettings:
         cls, backend: GroebnerBackend | str | None
     ) -> None:
         """Set the preferred method for performing Groebner Bases and related computations,
-        including :func:`.compute_groebner_basis`, :func:`.compute_primary_decomposition`, and :func:`.compute_radical`.
+        including :func:`.groebner.compute_groebner_basis`,
+        :func:`.groebner.compute_primary_decomposition`,
+        and :func:`.groebner.compute_radical`.
 
         Will default to singular.
